Got all model testing routines moved out of `reportlab_combinators.py`.
