+++ src/pyibisami/util/reportlab_combinators.py
@@ -8,7 +8,6 @@ Copyright (c) 2026 David Banas; All rights reserved World wide.
 """
 
 import platform
-import subprocess
 import sys
 
 from datetime import datetime
@@ -27,7 +26,8 @@ from ..ami.model import AMIModel
 from ..ami.parser import AMIParamConfigurator, ParamName
 from ..ibis.file import IBISModel
 from ..ibis.model import Model
-from ..testing.ami_tests import ami_tst_init_vs_getwave, ami_tst_samples_per_bit, ami_tst_getwave_input_length
+# from ..testing.ami_tests import ami_tst_init_vs_getwave, ami_tst_samples_per_bit, ami_tst_getwave_input_length
+# from ..testing.ibis_file_tests import get_ibis_contents, golden_parser_results
 
 from .tool_helpers import (
     init_vs_getwave, plot_sweeps, samples_per_bit, check_getwave_input_length,
@@ -69,283 +69,4 @@ def title_page(ibis_file: Path) -> list[Flowable]:
     flowables.append(Paragraph(f"{bold('NumPy:')} {np.__version__}", P))
     flowables.append(Paragraph(f"{bold('SciPy:')} {sp.__version__}", P))
     flowables.append(page_break)
-    return flowables
-
-
-# Model testing routines
-def get_ibis_contents(ibis_file: Path) -> tuple[IBISModel, list[Flowable]]:
-    """
-    List the components and models available in an IBIS model file.
-
-    Args:
-        ibis_file: The IBIS model file to be interrogated.
-
-    Returns:
-        A pair containing
-
-        - the ``IBISModel`` object, and
-        - the list of *ReportLab* ``Flowable``s describing the model contents.
-    """
-
-    flowables: list[Flowable] = [Paragraph("IBIS File Contents", H1)]
-
-    # Attempt to parse `*.ibs` file.
-    try:
-        ibis_model = IBISModel(str(ibis_file), False, gui=False)
-    except Exception as err:
-        raise RuntimeError(f"An error occurred while trying to read/parse the IBIS model file: {ibis_file}") from err
-
-    # Build the list of ReportLab Flowable elements for the IBIS model contents page of the report.
-    flowables.append(Paragraph(preformatted(ibis_model.info()), styles['Code']))
-    flowables.append(page_break)
-
-    return (ibis_model, flowables)
-
-
-def golden_parser_results(ibis_file: Path) -> list[Flowable]:
-    """
-    Run the ``ibischk`` parser on the IBIS file and return the results.
-
-    Args:
-        ibis_file: The IBIS model file to be interrogated.
-
-    Returns:
-        The list of *ReportLab* ``Flowable``s describing the parsing results.
-    """
-
-    flowables: list[Flowable] = [Paragraph("IBIS Golden Parser Results", H1)]
-    result = subprocess.run(['ibischk7_64', str(ibis_file)], capture_output=True, text=True)
-    flowables.append(Paragraph(preformatted(result.stdout), styles['Code']))
-    flowables.append(Paragraph(preformatted(result.stderr), styles['Code']))
-    flowables.append(page_break)
-    return flowables
-
-
-def test_ami_model(
-    model: Model,
-    ibis_file_dir: Path,
-    param_defs: list[TestSweep],
-    bit_rate: float,
-    nspui: int,
-    fig_x: float = 6,
-    fig_y: float = 3,
-) -> list[Flowable]:
-    """
-    Test an individual IBIS-AMI model.
-
-    Args:
-        model: The IBIS-AMI model to test.
-        ibis_file_dir: Parent directory of ``*.ibs`` file being tested.
-        param_defs: List of parameter definition sets to sweep over.
-        bit_rate: Bit rate to use for testing.
-        nspui: Number of samples per unit interval (a.k.a. - over-sampling factor).
-
-    Keyword Args:
-        fix_x: x-dimmension of plot (in.).
-            Default: 10
-        fix_y: y-dimmension of plot (in.).
-            Default: 3
-
-    Returns:
-        A list of *ReportLab* ``Flowable``s describing the test results.
-    """
-
-    if not model.is_ami:
-        return [Paragraph("Error: This model is not an IBIS-AMI model!", P)]
-
-    # Try to fetch AMI files for this machine/system combination.
-    machine = platform.machine().lower()
-    match(machine):
-        case "x86":
-            ami_dict = model.ami_files.get('32-bit')
-        case "x86_64":
-            ami_dict = model.ami_files.get('64-bit')
-        case "amd64":
-            ami_dict = model.ami_files.get('64-bit')
-        case "arm64":
-            ami_dict = model.ami_files.get('64-bit')
-        case _:
-            return [Paragraph(f"Error: Unrecognized machine type: {machine}!", P)]
-    system = platform.system().lower()
-    match(system):
-        case "windows":
-            ami_files = ami_dict.get('win')
-        case "linux":
-            ami_files = ami_dict.get('lin')
-        case "darwin":
-            ami_files = ami_dict.get('lin')
-        case _:
-            return [Paragraph(f"Error: Unrecognized system type: {system}!", P)]
-    if not ami_files:
-        return [Paragraph(f"Error: Model does not provide AMI files for this machine/system combination: {machine}/{system}!", P)]
-
-    flowables: list[Flowable] = [Paragraph("Basic Sanity Checking", H3)]
-
-    # Attempt to create the AMI model and its configurator.
-    dll_file, ami_file = list(map(lambda f: ibis_file_dir / Path(f), ami_files))
-    try:
-        ami_model = AMIModel(str(dll_file))
-    except Exception as err:
-        return [Paragraph(str(err), P), Paragraph(f"Error loading AMI DLL/SO: {dll_file}!", P)]
-    try:
-        with open(ami_file, mode="r", encoding="utf-8") as pfile:
-            pcfg = AMIParamConfigurator(pfile.read())
-        if pcfg.ami_parsing_errors:
-            flowables.append(
-                Paragraph(preformatted(f"Non-fatal AMI file parsing errors:\n{pcfg.ami_parsing_errors}"),
-                          P))
-        has_getwave = pcfg.fetch_param_val(["Reserved_Parameters", "GetWave_Exists"]) or False
-        ignore_bits = pcfg.fetch_param_val(["Reserved_Parameters", "Ignore_Bits"]) or 0
-        returns_impulse = pcfg.fetch_param_val(["Reserved_Parameters", "Init_Returns_Impulse"]) or False
-        has_ts4 = True if pcfg.fetch_param_val(["Reserved_Parameters", "Ts4file"]) else False
-        root_name = pcfg.input_ami_params[ParamName("root_name")]
-    except Exception as err:
-        flowables.append(Paragraph(str(err), P))
-        flowables.append(Paragraph(f"Error loading AMI parameter file: {ami_file}!", P))
-        return flowables
-
-    # Summarize results.
-    flowables.append(Paragraph("Model import was successful.", P))
-    if has_getwave:
-        flowables.append(Paragraph("Model has a `AMI_GetWave()` function.", P))
-        flowables.append(Paragraph(f"&nbsp;&nbsp;The first {ignore_bits} returned bits should be ignored.", P))
-    else:
-        flowables.append(Paragraph("Model has no `AMI_GetWave()` function.", P))
-    if returns_impulse:
-        flowables.append(Paragraph("Model's `AMI_Init()` function returns an impulse response.", P))
-    else:
-        flowables.append(Paragraph("Model's `AMI_Init()` function does not return an impulse response.", P))
-    if has_ts4:
-        flowables.append(Paragraph("Model includes on-die S-parameters.", P))
-    else:
-        flowables.append(Paragraph("Model does not include on-die S-parameters.", P))
-
-    # Test w/ perfect channel.
-    flowables.extend([
-        Paragraph("Testing w/ Perfect Channel", H3),
-        Paragraph("Here, we represent the 'channel' by a Kronecker delta function, \
-                  so as to elicit the actual responses of the model itself.", P),
-        spacer,
-        Paragraph("This imposes certain limitations on our testing. \
-                  For instance, since any equalization is over-equalization for such a perfect channel, \
-                  if the model imposes:"),
-        ListFlowable(
-            [Paragraph("a minimum amount of CTLE peaking, or"),
-             Paragraph("a minimum first DFE tap value,"),
-             Paragraph("etc.,"),
-            ], bulletType='bullet', bulletIndent=0.25*inch),
-        Paragraph("then we will end up in an over-equalized state."),
-        spacer,
-        Paragraph("Such a state will not usually affect the AMI_Init() function, \
-                  but may affect the AMI_GetWave() function, \
-                  particularly if a DFE has been enabled. \
-                  And this will cause disagreement between the two. Therefore:"),
-        spacer,
-        Paragraph("The results of this section should be considered with some skepticism, \
-                  as far as inferring how the model will behave in a 'real World' scenario.",
-                  bold_style),
-        spacer,
-        Paragraph(f"The {ital('Lossy Channel')} and {ital('Reflective Channel')} sections, below, \
-                  give a more useful assessment of the model's behavior in such real use scenarios."),
-    ])
-
-    bit_interval = 1.0 / bit_rate
-    sample_interval = bit_interval / nspui
-    channel_response = np.array([1.0] + [0.0] * (nspui - 1) + [0.0] * 19 * nspui) / sample_interval  # Kronecker delta
-
-    # - Init() vs. GetWave()
-    flowables.extend(
-        ami_tst_init_vs_getwave(
-            ami_model, pcfg, bit_interval, sample_interval,
-            channel_response, param_defs, fig_x=fig_x, fig_y=fig_y
-        )
-    )
-
-    # - samples per bit
-    flowables.extend(
-        ami_tst_samples_per_bit(
-            ami_model, pcfg, bit_interval, sample_interval,
-            channel_response, param_defs, fig_x=fig_x, fig_y=fig_y
-        )
-    )
-
-    # - GetWave() input length sensitivity
-    flowables.extend(
-        ami_tst_getwave_input_length(
-            ami_model, pcfg, bit_interval, sample_interval,
-            channel_response, param_defs, fig_x=fig_x, fig_y=fig_y
-        )
-    )
-
-    # Test w/ lossy channel.
-
-    # Test w/ reflective channel.
-
-    # Linearity check
-
-    return flowables
-
-
-def model_test_results(
-    ibis_file_dir: Path, ibis_model: IBISModel,
-    bit_rate: float, nspui: int,
-    param_defs: list[TestSweep],
-    model_name: Optional[str] = None,
-    debug: bool=False
-) -> list[Flowable]:
-    """
-    Test a subset of the IBIS-AMI models in the ``*.ibs`` file.
-
-    Args:
-        ibis_file_dir: The parent directory of the ``*.ibs`` file being tested.
-        ibis_model: The read and parsed ``*.ibs`` file.
-        bit_rate: The desired bit rate for testing.
-        nspui: Number of samples per unit interval (a.k.a. - over-sampling factor).
-        param_defs: The list of parameter sweep definitions to use for this test.
-
-    Keyword Args:
-        model_name: The name of a particular model to test.
-            Default = ``None`` (Means test all IBIS-AMI models found.)
-        debug: Include extra debugging output when ``True``.
-            Default = ``False``
-
-    Returns:
-        The list of *ReportLab* ``Flowable``s describing the testing results.
-    """
-
-    flowables: list[Flowable] = [Paragraph("IBIS-AMI Model(s) Testing Results", H1)]
-
-    if 'models' not in ibis_model.model_dict:
-        flowables.append(Paragraph("Error: The parsed IBIS file contained no model definitions!"))
-        flowables.append(page_break)
-        return flowables
-
-    def do_model(model_name: str) -> list[Flowable]:
-        """
-        Generate the output for one model.
-
-        Args:
-            model_name: Name of model to test.
-
-        Returns:
-            List of _ReportLab_ ``Flowable``s describing the model testing results.
-        """
-        flowables.append(Paragraph(f"Testing Model: {model_name}", H2))
-        model = ibis_model.model_dict['models'][model_name]
-        flowables.append(Paragraph(preformatted(f"{model}"), styles['Code']))
-        flowables.extend(test_ami_model(model, ibis_file_dir, param_defs, bit_rate, nspui))
-        flowables.append(page_break)
-        return flowables
-
-    if model_name:
-        if model_name not in ibis_model.model_dict['models']:
-            flowables.append(Paragraph(f"Error: The requested model: {model_name}, was not found!"))
-            flowables.append(page_break)
-            return flowables
-        do_model(model_name)
-        return flowables
-
-    for model_name in ibis_model.model_dict['models']:
-        do_model(model_name)
-
     return flowables
